requirements.txt
pdfplumber
pytesseract
Pillow
psutil
pandas
pdf2image